+++ src/components/DatasetSelect.tsx
@@ -4,10 +4,13 @@
  * https://opensource.org/licenses/MIT.
  */
 
+import { ReactElement, useMemo } from "react";
+import Divider from "@mui/material/Divider";
 import Input from "@mui/material/Input";
 import InputLabel from "@mui/material/InputLabel";
 import MenuItem from "@mui/material/MenuItem";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
+import Typography from "@mui/material/Typography";
 import TravelExploreIcon from "@mui/icons-material/TravelExplore";
 
 import i18n from "@/i18n";
@@ -15,25 +18,28 @@ import { Dataset } from "@/model/dataset";
 import { WithLocale } from "@/util/lang";
 import ToolButton from "@/components/ToolButton";
 import ControlBarItem from "./ControlBarItem";
-import { ReactElement, useMemo } from "react";
-import Divider from "@mui/material/Divider";
-import Typography from "@mui/material/Typography";
+import { LayerVisibilities } from "@/states/controlState";
+import { commonStyles } from "@/components/common-styles";
 
 interface DatasetSelectProps extends WithLocale {
-  selectedDatasetId: string | null;
+  selectedDataset: Dataset | null;
   datasets: Dataset[];
   selectDataset: (
     datasetId: string | null,
     datasets: Dataset[],
     showInMap: boolean,
   ) => void;
+  layerVisibilities: LayerVisibilities;
+  toggleDatasetRgbLayer: (visible: boolean) => void;
   locateSelectedDataset: () => void;
 }
 
 export default function DatasetSelect({
-  selectedDatasetId,
+  selectedDataset,
   datasets,
   selectDataset,
+  layerVisibilities,
+  toggleDatasetRgbLayer,
   locateSelectedDataset,
 }: DatasetSelectProps) {
   const sortedDatasets = useMemo(() => {
@@ -69,7 +75,7 @@ export default function DatasetSelect({
     selectDataset(datasetId, datasets, true);
   };
 
-  selectedDatasetId = selectedDatasetId || "";
+  const selectedDatasetId = selectedDataset ? selectedDataset.id : "";
   datasets = datasets || [];
 
   const datasetSelectLabel = (
@@ -119,19 +125,40 @@ export default function DatasetSelect({
     </Select>
   );
 
-  const locateDatasetButton = (
-    <ToolButton
-      onClick={locateSelectedDataset}
-      tooltipText={i18n.get("Locate dataset in map")}
-      icon={<TravelExploreIcon />}
-    />
-  );
+  const rgbVisible =
+    layerVisibilities.datasetRgb && !layerVisibilities.datasetVariable;
 
   return (
     <ControlBarItem
       label={datasetSelectLabel}
       control={datasetSelect}
-      actions={locateDatasetButton}
+      actions={
+        <>
+          {selectedDataset?.rgbSchema && (
+            <ToolButton
+              onClick={() => toggleDatasetRgbLayer(!rgbVisible)}
+              tooltipText={i18n.get(
+                "Switch between dataset RGB layer and variable layer",
+              )}
+              sx={{
+                ...commonStyles.toggleButton,
+                width: "26.42px",
+                height: "26.42px",
+                marginLeft: 0.4,
+              }}
+              icon={<Typography fontSize={9}>RGB</Typography>}
+              toggle
+              value="rgb"
+              selected={rgbVisible}
+            />
+          )}
+          <ToolButton
+            onClick={locateSelectedDataset}
+            tooltipText={i18n.get("Locate dataset in map")}
+            icon={<TravelExploreIcon />}
+          />
+        </>
+      }
       sx={{ marginLeft: 0 }}
     />
   );